// backend/src/modules/reservations/enums/reservation-error.enum.ts
export enum ReservationError {
  SEAT_NOT_FOUND = 'SEAT_NOT_FOUND',
  SEAT_NOT_AVAILABLE = 'SEAT_NOT_AVAILABLE',
  RESERVATION_NOT_FOUND = 'RESERVATION_NOT_FOUND',
  RESERVATION_NOT_YOURS = 'RESERVATION_NOT_YOURS',
  RESERVATION_CANNOT_BE_CONFIRMED = 'RESERVATION_CANNOT_BE_CONFIRMED',
  RESERVATION_EXPIRED = 'RESERVATION_EXPIRED',
}